requirements.txt
python-dotenv
typing
psycopg2-binary
SQLAlchemy
uvicorn
pydantic
"fastapi[standard]"